server/requirements.txt
geoip2==2.9.0
Django==2.1.7
djangorestframework==3.9.2
